requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
cssselect==1.2.0
d==0.2.2
Django==4.2.4
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
lxml==4.9.3
Markdown==3.4.4
Pygments==2.16.1
pyquery==2.0.0
pytz==2023.3.post1
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
